Move rendering.py to utils
