+++ gym/envs/classic_control/acrobot.py
@@ -182,7 +182,7 @@ class AcrobotEnv(core.Env):
         return (dtheta1, dtheta2, ddtheta1, ddtheta2, 0.0)
 
     def render(self, mode="human"):
-        from gym.envs.classic_control import rendering
+        from gym.utils import rendering
 
         s = self.state
