+++ gym/envs/classic_control/continuous_mountain_car.py
@@ -132,7 +132,7 @@ class Continuous_MountainCarEnv(gym.Env):
         carheight = 20
 
         if self.viewer is None:
-            from gym.envs.classic_control import rendering
+            from gym.utils import rendering
 
             self.viewer = rendering.Viewer(screen_width, screen_height)
             xs = np.linspace(self.min_position, self.max_position, 100)
