+++ gym/envs/classic_control/mountain_car.py
@@ -110,7 +110,7 @@ class MountainCarEnv(gym.Env):
         carheight = 20
 
         if self.viewer is None:
-            from gym.envs.classic_control import rendering
+            from gym.utils import rendering
 
             self.viewer = rendering.Viewer(screen_width, screen_height)
             xs = np.linspace(self.min_position, self.max_position, 100)
