+++ gym/envs/unittest/cube_crash.py
@@ -147,7 +147,7 @@ class CubeCrash(gym.Env):
             return self.last_obs
 
         elif mode == "human":
-            from gym.envs.classic_control import rendering
+            from gym.utils import rendering
 
             if self.viewer is None:
                 self.viewer = rendering.SimpleImageViewer()
